requirements.txt
Django
djangorestframework
django-filter
djangorestframework-recursive
django-mptt